Add preoption param to reuse _run_ffmpeg function

+++ lib/video_editor/ffmpeg.py
@@ -1,9 +1,11 @@
 import json
-import os
-import subprocess
 import logging
+import os
+import shlex
+import subprocess
 
 from flask import current_app as app
+
 from lib.utils import create_temp_file
 
 from .interface import VideoEditorInterface
@@ -130,7 +132,7 @@ class FFMPEGVideoEditor(VideoEditorInterface):
                 os.remove(path_input)
         return content, metadata_edit_file
 
-    def capture_thumbnail(self, stream_file, filename, duration, position, crop={}, rotate=0):
+    def capture_thumbnail(self, stream_file, filename, duration, position, crop=None, rotate=0):
         """
         Use ffmpeg tool to capture video frame at a position.
         :param stream_file: video file
@@ -156,7 +158,6 @@ class FFMPEGVideoEditor(VideoEditorInterface):
                 position = duration - 0.1
             # create output file path
             output_file = f"{path_video}_preview_thumbnail.png"
-            cmd = ["-accurate_seek", "-i", path_video, "-ss", str(position), "-vframes", "1"]
 
             vfilter = ''
             if crop:
@@ -165,12 +166,19 @@ class FFMPEGVideoEditor(VideoEditorInterface):
                 vfilter += ',' if vfilter else '-vf '
                 transpose = f'transpose=1' if rotate > 0 else f'transpose=2'
                 vfilter += ','.join([transpose] * (rotate // 90))
-                print(vfilter)
-            if vfilter:
-                cmd.extend(vfilter.split(' '))
 
             # run ffmpeg command
-            subprocess.run(["ffmpeg", "-v", "error", "-y", *cmd, output_file])
+            self._run_ffmpeg(
+                path_input=path_video,
+                path_output=output_file,
+                preoptions=('-y', '-accurate_seek'),
+                options=(
+                    '-ss', str(position),
+                    '-vframes', '1',
+                    *shlex.split(vfilter),
+                ),
+                override=False,
+            )
             try:
                 # get metadata
                 thumbnail_metadata = self._get_meta(output_file)
@@ -215,7 +223,7 @@ class FFMPEGVideoEditor(VideoEditorInterface):
             # subprocess bash -> ffmpeg in the loop
             subprocess.run([path_script, path_video, output_file, str(frame_per_second), str(thumbnails_amount)])
             for i in range(0, thumbnails_amount):
-                thumbnail_path =  f'{output_file}{i}.png'
+                thumbnail_path = f'{output_file}{i}.png'
                 try:
                     # get metadata
                     thumbnail_metadata = self._get_meta(thumbnail_path)
@@ -230,27 +238,31 @@ class FFMPEGVideoEditor(VideoEditorInterface):
         finally:
             os.remove(path_video)
 
-    def _run_ffmpeg(self, path_input, path_output, options=tuple()):
+    def _run_ffmpeg(self, path_input, path_output, preoptions=tuple(), options=tuple(), override=True):
         """
         Subprocess `ffmpeg` command.
         :param path_input: input file path
         :type path_input: str
         :param path_output: outut file path
         :type path_output: str
+        :param preoptions: options apply for input file
+        :type preoptions: tuple
         :param options: options for ffmpeg cmd
         :type options: tuple
+        :param override: replace input file with output file
+        :type override: bool
         :return: file path to edited file
         :rtype: str
         """
-        try:
-            # run ffmpeg with provided options
-            subprocess.run(["ffmpeg", "-loglevel", "error", "-i", path_input, *options, path_output])
-            # replace tmp origin
-            subprocess.run(["cp", "-r", path_output, path_input])
-            return path_input
-        finally:
-            # delete old tmp input file
-            os.remove(path_output)
+        # run ffmpeg with provided options
+        subprocess.run(["ffmpeg", "-loglevel", "error", *preoptions, "-i", path_input, *options, path_output])
+        if not override:
+            return path_output
+        # replace tmp origin
+        subprocess.run(["cp", "-r", path_output, path_input])
+        # delete old tmp input file
+        os.remove(path_output)
+        return path_input
 
     def _get_meta(self, file_path):
         """
